+++ packages/dashboard/src/components/Wizard/SetupWizard.styles.ts
@@ -1,29 +1,29 @@
 import { ISetupWizardStyles } from './SetupWizard.types';
-import { DefaultFontStyles } from 'office-ui-fabric-react/lib/Styling';
-
-const regularFontSize = 14;
-const regularLineHeight = 20;
-const fontWeight = DefaultFontStyles.smallPlus.fontWeight;
-const commonFontFamily = DefaultFontStyles.medium.fontFamily;
-const subwayNavBaseColor = '#333333';
+import { DefaultFontStyles } from 'office-ui-fabric-react';
+import { NeutralColors } from '@uifabric/fluent-theme';
 
 export const getSetupWizardStyles = (): ISetupWizardStyles => {
   return {
-    wizardContainer: {
-      fontSize: regularFontSize,
-      lineHeight: regularLineHeight,
-      fontWeight: fontWeight,
-      fontFamily: commonFontFamily,
-      color: subwayNavBaseColor,
-      display: 'flex',
-      flexDirection: 'column'
-    },
+    wizardContainer: [
+      DefaultFontStyles.medium,
+      {
+        color: NeutralColors.gray130,
+        display: 'flex',
+        flexDirection: 'column'
+      }
+    ],
     titleSection: {
-      border: '1px solid #E1DFDD',
+      border: `1px solid ${NeutralColors.gray40}`,
       height: '75px'
     },
-    title: {
-      margin: '30px'
+    actionBarSection: {
+      whiteSpace: 'nowrap',
+      width: '100%',
+      display: 'flex',
+      flexDirection: 'row',
+      border: `1px solid ${NeutralColors.gray40}`,
+      height: '75px',
+      alignItems: 'center'
     }
   };
 };