+++ packages/dashboard/src/components/Wizard/SetupWizard.types.ts
@@ -1,12 +1,15 @@
-import { IStyle } from 'office-ui-fabric-react/lib/Styling';
-import { IWizardProps, IWizardStepAction } from './Wizard.types';
+import { IStyle, ITheme } from 'office-ui-fabric-react/lib/Styling';
+import { IWizardProps } from './Wizard.types';
 
-export interface ISetupWizardProps extends IWizardProps {
+export interface ISetupWizardProps {
   wizardTitle?: ISetupWizardTitleProps;
 
-  exitWizardAction: IWizardStepAction;
+  /** pass-thru wizard props */
+  wizardProps: IWizardProps;
 
-  backClickAction: IWizardStepAction;
+  styles?: ISetupWizardStyles;
+
+  theme?: ITheme;
 }
 
 export interface ISetupWizardTitleProps {
@@ -19,5 +22,9 @@ export interface ISetupWizardStyles {
 
   titleSection: IStyle;
 
-  title: IStyle;
+  actionBarSection: IStyle;
+}
+
+export interface ISetupWizardStyleProps {
+  theme: ITheme;
 }