+++ packages/dashboard/src/components/Wizard/Wizard.types.ts
@@ -1,5 +1,5 @@
-import { IStyle } from 'office-ui-fabric-react/lib/Styling';
-import { ISubwayNavStep, SubwayNavStepState } from '../SubwayNav/SubwayNav.types';
+import { IStyle, ITheme } from 'office-ui-fabric-react/lib/Styling';
+import { ISubwayNavNodeProps, SubwayNavNodeState } from '../SubwayNav/SubwayNode.types';
 
 export interface IWizardProps {
   // List of steps in the wizard
@@ -7,33 +7,48 @@ export interface IWizardProps {
 
   wizardComplete?: boolean;
 
-  allowSkipAhead?: boolean;
+  /* Step that is not part of navigation, only shown when wizard is complete */
+  wizardCompleteStep?: IWizardStepProps;
+
+  stepToShow?: IWizardStepProps;
+
+  styles?: IWizardStyles;
+
+  theme?: ITheme;
 }
 
 export interface IWizardStepProps {
-  key: string;
+  id: string;
+
+  /**
+   * Optional ID for the parent of the step.
+   * to aid in data operations
+   */
+  parentId?: string;
 
   label: string;
 
-  state: SubwayNavStepState;
+  state: SubwayNavNodeState;
 
   disabled?: boolean;
 
   isSubStep?: boolean;
 
-  onClickStep: (step: ISubwayNavStep, subStep: ISubwayNavStep | undefined) => void;
+  onClickStep?: (step: ISubwayNavNodeProps) => void;
 
   wizardContent?: IWizardContentProps;
 
   subSteps?: IWizardStepProps[];
+
+  titleElement: JSX.Element;
+
+  footerElement: JSX.Element;
 }
 
 export interface IWizardContentProps {
-  contentTitle?: string;
+  contentTitleElement?: JSX.Element;
 
   content: JSX.Element;
-
-  mainAction: IWizardStepAction;
 }
 
 export interface IWizardStepAction {
@@ -45,6 +60,10 @@ export interface IWizardStepAction {
 
   // Action is disabled or not
   disabled?: boolean;
+
+  className?: string;
+
+  currentStep?: IWizardStepProps;
 }
 
 export interface IWizardTitleProps {
@@ -67,5 +86,11 @@ export interface IWizardStyles {
 
   contentSection: IStyle;
 
+  contentTitle: IStyle;
+
   content: IStyle;
+}
+
+export interface IWizardStyleProps {
+  theme: ITheme;
 }