+++ packages/fluent-theme/src/fluent/styles/Rating.styles.ts
@@ -1,9 +1,8 @@
 import { IRatingStyleProps, IRatingStyles } from 'office-ui-fabric-react/lib/Rating';
-import { NeutralColors } from '../FluentColors';
 
 export const RatingStyles = (props: IRatingStyleProps): Partial<IRatingStyles> => {
   const { disabled, readOnly, theme } = props;
-  const { palette, semanticColors } = theme;
+  const { palette } = theme;
 
   return {
     root: [
@@ -19,30 +18,8 @@ export const RatingStyles = (props: IRatingStyleProps): Partial<IRatingStyles> =
           }
         }
     ],
-    ratingStarBack: [
-      {
-        color: NeutralColors.gray80
-      },
-      disabled && {
-        color: semanticColors.disabledBodySubtext
-      }
-    ],
     ratingStarFront: {
       color: palette.neutralPrimary
-    },
-    ratingButton: [
-      !disabled &&
-        !readOnly && {
-          selectors: {
-            // This is part 2 of highlighting all stars up to the one the user is hovering over
-            '&:hover ~ .ms-Rating-button': {
-              selectors: {
-                '.ms-RatingStar-back': { color: NeutralColors.gray80 },
-                '.ms-RatingStar-front': { color: NeutralColors.gray80 }
-              }
-            }
-          }
-        }
-    ]
+    }
   };
 };